build: restructured files with automatic versioning
