+++ .gitignore
@@ -1,3 +1,6 @@
+# Custom ignores
+pombseen/_version.py
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
