--- README.md
@@ -1,5 +1,0 @@
-pomBseen segmentation for Cell-ACDC.
-
-[Cell-ACDC publication](https://bmcbiol.biomedcentral.com/articles/10.1186/s12915-022-01372-6)
-[Cell-ADCD github repo](https://github.com/SchmollerLab/Cell_ACDC)
-[pomBseen publication](https://pubmed.ncbi.nlm.nih.gov/37699057/)
