+++ README.rst
@@ -1,0 +1,5 @@
+pomBseen
+========
+
+Python implementation of `pomBseen <https://pubmed.ncbi.nlm.nih.gov/37699057>`_:sup:`1`.
+This model is available to use in `Cell-ACDC <https://github.com/SchmollerLab/Cell_ACDC>`_.
