+++ pombseen/__init__.py
@@ -1,0 +1,10 @@
+try:
+    from setuptools_scm import get_version
+    __version__ = get_version(root='..', relative_to=__file__)
+except Exception as e:
+    try:
+        from ._version import version as __version__
+    except ImportError:
+        __version__ = "not-installed"
+
+__author__ = 'Timon Stegmaier and Francesco Padovani'
